packages/image-alignment-component/pyproject.toml
[project]
name = "image-alignment-component"
version = "0.1.0"
description = "A custom Gradio component for aligning two images with keyboard controls"
readme = "README.md"
authors = [
    { name = "Emia", email = "47446852+YuShigurey@users.noreply.github.com" }
]
requires-python = ">=3.9,<3.13"
dependencies = [
    "gradio>=3.50.0",
    "pillow>=8.0.0",
    "numpy>=1.20.0",
    "matplotlib>=3.5.0",
    "PyQt6>=6.6.0",
    "pycocotools>=2.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
